# client/src/pages/Swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CloudXi Museum",
    "description": "This is the REST API documentaion for the Cloudxi Museum online management system",
    "version": "1.0.0",
    "contact": {
      "email": "CloudXiMuseum@gmail.com"
    }
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
...
  "paths": {
...
    "/default/artworks/{artID}": {
      "get": {
        "tags": [
          "Gallery & Artwork"
        ],
        "summary": "Retrieves information of the selected artwork",
        "description": "Retrieves information of the selected artwork",
        "parameters": [
          {
            "name": "artID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/customer/user-likes/{artID}/like": {
      "post": {
        "tags": [
          "Gallery & Artwork"
        ],
        "summary": "Give like for the picture",
        "description": "Give like for the picture",
        "parameters": [
          {
            "name": "artID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "like": {
                  "example": "any"
                },
                "UID": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/customer/user-likes/{artID}/{userID}": {
      "get": {
        "tags": [
          "Gallery & Artwork"
        ],
        "summary": "Check if the user has liked the picture or not",
        "description": "Check if the user has liked the picture or not",
        "parameters": [
          {
            "name": "artID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "userID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/customer/reviews/{uid}": {
      "get": {
        "tags": [
          "Review"
        ],
        "summary": "Get the reviews of the customer",
        "description": "Get the reviews of the customer",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/customer/reviews/": {
      "post": {
        "tags": [
          "Review"
        ],
        "summary": "Give reviews for the Museum",
        "description": "Give reviews for the Museum",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/employee/modify/{tableName}/insert": {
      "post": {
        "tags": [
          "Modify"
        ],
        "summary": "Inserts a new tuple to a table",
        "description": "Inserts a new tuple to a table",
        "parameters": [
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/employee/modify/{tableName}/delete": {
      "post": {
        "tags": [
          "Modify"
        ],
        "summary": "Delete a tuple from a table",
        "description": "Delete a tuple from a table",
        "parameters": [
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/employee/modify/{tableName}/modify": {
      "post": {
        "tags": [
          "Modify"
        ],
        "summary": "Modify a tuple in a table",
        "description": "Modify a tuple in a table",
        "parameters": [
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/admin/modify/{tableName}/insert": {
      "post": {
        "tags": [
          "Modify"
        ],
        "summary": "Inserts a new tuple to a table",
        "description": "Inserts a new tuple to a table",
        "parameters": [
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/admin/modify/{tableName}/delete": {
      "post": {
        "tags": [
          "Modify"
        ],
        "summary": "Delete a tuple from a table",
        "description": "Delete a tuple from a table",
        "parameters": [
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/admin/modify/{tableName}/modify": {
      "post": {
        "tags": [
          "Modify"
        ],
        "summary": "Modify a tuple in a table",
        "description": "Modify a tuple in a table",
        "parameters": [
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  }
}